Change sum type from double to short in TssCalculator
Updated the sum variable in TssCalculator from double to short,
as power values are within a range suitable for short. This
improves efficiency and better matches the expected data range.

+++ src/FitSyncHub.Common/TssCalculator.cs
@@ -55,7 +55,7 @@ public static class TssCalculator
 
         List<double> rollingAverages = [];
         // short should be enough, power values are usually within 0-2000 range
-        double sum = 0;
+        short sum = 0;
 
         // Initial window
         for (var i = 0; i < WindowSize; i++)
